requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
firebase-admin==5.2.0
frozenlist==1.3.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
msgpack==1.0.4
multidict==6.0.2
proto-plus==1.20.6
protobuf==3.20.1
py-cord==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.11
yarl==1.8.0
